SuggestionBox/SuggestBox を SearchablePopupMenu に rename
