+++ test/components/SearchablePopupMenu.test.tsx
@@ -1,7 +1,7 @@
 import { act, fireEvent, render } from '@testing-library/preact';
 import userEvent from '@testing-library/user-event';
 import { datatype } from 'faker';
-import { SuggestionBox } from '../../src/components/SuggestionBox';
+import { SearchablePopupMenu } from '../../src/components/SearchablePopupMenu';
 import { forwardMatcher } from '../../src/lib/matcher';
 import { CursorPosition } from '../../src/types';
 import { keydownEnterEvent, keydownEscapeEvent } from '../helpers/key';
@@ -17,10 +17,10 @@ const items = [
 const matcher = forwardMatcher;
 const props = { cursorPosition, items, matcher };
 
-describe('SuggestionBox', () => {
+describe('SearchablePopupMenu', () => {
   describe('open === false の時', () => {
     test('ポップアップも QueryInput も表示されない', () => {
-      const { asFragment, queryByTestId } = render(<SuggestionBox open={false} {...props} />);
+      const { asFragment, queryByTestId } = render(<SearchablePopupMenu open={false} {...props} />);
       expect(queryByTestId('popup-menu')).toBeNull();
       expect(queryByTestId('query-input')).toBeNull();
       expect(asFragment()).toMatchSnapshot();
@@ -28,14 +28,14 @@ describe('SuggestionBox', () => {
   });
   describe('open === true の時', () => {
     test('ポップアップと QueryInput が表示される', () => {
-      const { asFragment, queryByTestId } = render(<SuggestionBox open {...props} />);
+      const { asFragment, queryByTestId } = render(<SearchablePopupMenu open {...props} />);
       expect(queryByTestId('popup-menu')).not.toBeNull();
       expect(queryByTestId('query-input')).not.toBeNull();
       expect(asFragment()).toMatchSnapshot();
     });
     test('Esc 押下で onClose が呼び出される', async () => {
       const onClose = jest.fn();
-      render(<SuggestionBox open {...props} onClose={onClose} />);
+      render(<SearchablePopupMenu open {...props} onClose={onClose} />);
       expect(onClose).toBeCalledTimes(0);
       await act(() => {
         fireEvent(document, keydownEscapeEvent);
@@ -45,13 +45,13 @@ describe('SuggestionBox', () => {
     describe('ポップアップに表示されるアイテムが空の時', () => {
       test('emptyMessage でアイテムが空の時のメッセージを変更できる', () => {
         const emptyMessage = datatype.string();
-        const { getByText } = render(<SuggestionBox open {...props} items={[]} emptyMessage={emptyMessage} />);
+        const { getByText } = render(<SearchablePopupMenu open {...props} items={[]} emptyMessage={emptyMessage} />);
         expect(getByText(emptyMessage)).toBeInTheDocument();
       });
     });
     describe('ポップアップに表示されるアイテムが空でない時', () => {
       test('QueryInput に文字を入力するとアイテムがフィルタされる', () => {
-        const { getByTestId } = render(<SuggestionBox open {...props} />);
+        const { getByTestId } = render(<SearchablePopupMenu open {...props} />);
         const buttonContainer = getByTestId('button-container');
         const queryInput = getByTestId('query-input');
 
@@ -67,7 +67,7 @@ describe('SuggestionBox', () => {
       });
       test('Enter 押下で onSelect が呼び出される', async () => {
         const onSelect = jest.fn();
-        render(<SuggestionBox open {...props} onSelect={onSelect} />);
+        render(<SearchablePopupMenu open {...props} onSelect={onSelect} />);
         expect(onSelect).toBeCalledTimes(0);
         await act(() => {
           fireEvent(document, keydownEnterEvent);
@@ -77,13 +77,13 @@ describe('SuggestionBox', () => {
     });
   });
   test('open === true になった時に、 QueryInput に入力された文字がリセットされる', () => {
-    const { rerender, getByTestId } = render(<SuggestionBox open {...props} />);
+    const { rerender, getByTestId } = render(<SearchablePopupMenu open {...props} />);
 
     userEvent.type(getByTestId('query-input'), 'a');
     expect(getByTestId('query-input')).toHaveValue('a');
 
-    rerender(<SuggestionBox open={false} {...props} />);
-    rerender(<SuggestionBox open {...props} />);
+    rerender(<SearchablePopupMenu open={false} {...props} />);
+    rerender(<SearchablePopupMenu open {...props} />);
 
     expect(getByTestId('query-input')).toHaveValue('');
   });
